docker_image_raw/Queries/python/requirements.txt
psycopg2-binary>=2.9.9
pandas>=2.0.0
matplotlib>=3.7.0
jupyter>=1.0.0
notebook>=7.0.0 